fallbackで置換をする
