+++ help/CompileChm.ps1
@@ -3,7 +3,8 @@ Param(
 	[Parameter(Mandatory)]
     [string]$HtmlHelpProject,
 	[Parameter(Mandatory)]
-    [string]$Destination
+    [string]$Destination,
+    $CMD_HHC
 )
 
 # ファイル存在チェック
@@ -16,7 +17,7 @@ function Is-Existing
 		[string]$Path
 	)
 
-    return Test-Path -Path $Path
+    return Test-Path -LiteralPath $Path
 }
 
 # ファイル不存在チェック
@@ -91,7 +92,7 @@ function Copy-Chm
 	{
 		if ($Destination | Is-Missing)
 		{
-			New-Item -Path $Destination -ItemType Directory
+			New-Item -LiteralPath $Destination -ItemType Directory
 		}
 
 		$Destination = [System.IO.Path]::Combine($Destination, [System.IO.Path]::GetFileName($Path))
@@ -107,7 +108,7 @@ function Copy-Chm
 	{
 		echo "`$CompiledHelp is: $Path"
 		echo "`$Destination  is: $Destination"
-		Copy-Item -Path $Path -Destination $Destination
+		Copy-Item -LiteralPath $Path -Destination $Destination
 	}
 	catch #[System.IO.IOException]
 	{
@@ -118,35 +119,34 @@ function Copy-Chm
 	return $true
 }
 
-$HtmlHelpProject = Convert-Path $HtmlHelpProject
-
+$HtmlHelpProject = Convert-Path -LiteralPath $HtmlHelpProject
 if (-not($HtmlHelpProject -imatch '\.hhp$'))
 {
 	throw [System.ArgumentException]::new("Bad filename of `$HtmlHelpProject: $HtmlHelpProject")
 }
 
-if ([string]::IsNullOrEmpty($env:CMD_HHC) -or ($env:CMD_HHC | Is-Missing))
+if ([string]::IsNullOrEmpty($CMD_HHC) -or ($CMD_HHC | Is-Missing))
 {
-	$env:CMD_HHC = 'C:\Program Files (x86)\HTML Help Workshop\hhc.exe'
+	$CMD_HHC = 'C:\Program Files (x86)\HTML Help Workshop\hhc.exe'
 }
 
-$hhc = $env:CMD_HHC -replace '(.+)', '""$1""'
+$hhc = $CMD_HHC -replace '(.+)', '""$1""'
 $CompiledHelp = $HtmlHelpProject -ireplace '\.hhp$', '.chm'
 $CompileLog = "$([System.IO.Path]::GetDirectoryName($HtmlHelpProject))\Compile.Log"
 
 if ($CompileLog | Is-Existing)
 {
-	rm $CompileLog
+	rm -LiteralPath $CompileLog
 }
 
 if ($CompiledHelp | Is-Existing)
 {
-	rm $CompiledHelp
+	rm -LiteralPath $CompiledHelp
 }
 
 if ($Destination | Is-Existing)
 {
-	rm $Destination
+	rm -LiteralPath $Destination
 }
 
 while ($true)
@@ -185,6 +185,7 @@ while ($true)
 			}
 			elseif (Copy-Chm $CompiledHelp $Destination)
 			{
+				echo "Copy `$CompiledHelp is done"
 				return
 			}
 			else
