+++ help/EncoderEscapingFallback.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
 	Copyright (C) 2018-2022, Sakura Editor Organization
 
 	This software is provided 'as-is', without any express or implied
@@ -24,49 +24,145 @@
 using System;
 using System.Text;
 using System.Text.RegularExpressions;
+using System.Security;
+using System.Linq;
 
-namespace ChmSourceConverter
+/// <summary>
+/// Provides a failure handling mechanism, called a fallback,
+/// for an input character that cannot be converted to an output byte sequence.
+/// The fallback uses a user-specified replacement string instead of the original input character.
+/// This class cannot be inherited.
+/// </summary>
+public sealed class EncoderEscapingFallback : EncoderFallback
 {
     /// <summary>
-    /// Provides a failure handling mechanism, called a fallback,
-    /// for an input character that cannot be converted to an output byte sequence.
-    /// The fallback uses a user-specified replacement string instead of the original input character.
-    /// This class cannot be inherited.
+    /// the Escaping Format
     /// </summary>
-    public sealed class EncoderEscapingFallback : EncoderFallback
+    public string Format { get; private set; }
+
+    /// <summary>
+    /// construct the instance
+    /// </summary>
+    /// <param name="format">escaping format, should be including "{0}".</param>
+    public EncoderEscapingFallback(string format)
+        : base()
     {
-        /// <summary>
-        /// the Escaping Format
-        /// </summary>
-        public string Format { get; private set; }
+        // validate format string.
+        if (format == null)
+            throw new ArgumentNullException("format");
+        if (!Regex.IsMatch(format, @".*\{0.*\}.*"))
+            throw new ArgumentException("bad format", "format");
 
-        /// <summary>
-        /// construct the instance
-        /// </summary>
-        /// <param name="format">escaping format, should be including "{0}".</param>
-        public EncoderEscapingFallback(string format)
-            : base()
+        // test formatting
+        string.Format(format, 1);
+
+        // set internal member.
+        Format = format;
+    }
+
+    /// <inheritdoc />
+    public override int MaxCharCount { get{ return EncoderFallback.ExceptionFallback.MaxCharCount; }}
+
+    /// <inheritdoc />
+    public override EncoderFallbackBuffer CreateFallbackBuffer()
+    {
+        return new EncoderEscapingFallbackBuffer(this);
+    }
+}
+/// <summary>
+/// Represents a substitute input string that is used when the original input character cannot be encoded.
+/// This class cannot be inherited.
+/// </summary>
+public sealed class EncoderEscapingFallbackBuffer : EncoderFallbackBuffer
+{
+    private ArraySegment<char> Buffer;
+    private EncoderEscapingFallback FallbackRef;
+
+    /// <inheritdoc />
+    public override int Remaining { get{ return Buffer.Count(); }}
+
+    /// <summary>
+    /// Initializes a new instance of the EncoderEscapingFallbackBuffer class
+    /// using the value of a EncoderEscapingFallback object.
+    /// </summary>
+    /// <param name="fallback">A EncoderEscapingFallback object.</param>
+    public EncoderEscapingFallbackBuffer(EncoderEscapingFallback fallback)
+    {
+        Buffer = new ArraySegment<char>(Array.Empty<char>());
+        FallbackRef = fallback;
+    }
+
+    /// <summary>
+    /// Prepares the escaping fallback buffer to use the current format string.
+    /// </summary>
+    /// <param name="charUnknown"></param>
+    /// <returns></returns>
+    private bool Fallback(int charUnknown)
+    {
+        var escapedChar = string.Format(FallbackRef.Format, charUnknown);
+        Buffer = new ArraySegment<char>(escapedChar.ToCharArray());
+        return true;
+    }
+
+    /// <inheritdoc />
+    public override bool Fallback(char charUnknown, int index)
+    {
+        if (Buffer.Any())
         {
-            // validate format string.
-            if (format == null)
-                throw new ArgumentNullException("format");
-            if (!Regex.IsMatch(format, @".*\{0.*\}.*"))
-                throw new ArgumentException("bad format", "format");
-
-            // test formatting
-            string.Format(format, 1);
-
-            // set internal member.
-            Format = format;
+            throw new ArgumentException("This method is called again before the GetNextChar method has read all the replacement string characters.");
         }
 
-        /// <inheritdoc />
-        public override int MaxCharCount => EncoderFallback.ExceptionFallback.MaxCharCount;
+        return Fallback((Int32)charUnknown);
+    }
 
-        /// <inheritdoc />
-        public override EncoderFallbackBuffer CreateFallbackBuffer()
+    /// <inheritdoc />
+    public override bool Fallback(char charUnknownHigh, char charUnknownLow, int index)
+    {
+        if (Buffer.Any())
         {
-            return new EncoderEscapingFallbackBuffer(this);
+            throw new ArgumentException("This method is called again before the GetNextChar method has read all the replacement string characters.");
+        }
+        if (charUnknownHigh < 0xD800 || charUnknownHigh > 0xD8FF)
+        {
+            throw new ArgumentOutOfRangeException("charUnknownHigh", charUnknownHigh, "The value of charUnknownHigh is out of range");
+        }
+        if (charUnknownLow < 0xDC00 || charUnknownLow > 0xDFFF)
+        {
+            throw new ArgumentOutOfRangeException("charUnknownLow", charUnknownLow, "The value of charUnknownLow is out of range");
+        }
+
+        int charUnknown = Char.ConvertToUtf32(charUnknownHigh, charUnknownLow);
+        return Fallback(charUnknown);
+    }
+
+    /// <inheritdoc />
+    public override char GetNextChar()
+    {
+        char ch = Buffer.FirstOrDefault();
+        if (Buffer.Any())
+        {
+            Buffer = new ArraySegment<char>(Buffer.Array, Buffer.Offset + 1, Buffer.Count - 1);
+        }
+        return ch;
+    }
+
+    /// <inheritdoc />
+    public override bool MovePrevious()
+    {
+        if (Buffer.Offset == 0) return false;
+        Buffer = new ArraySegment<char>(Buffer.Array, Buffer.Offset - 1, Buffer.Count + 1);
+        return true;
+    }
+
+    /// <summary>
+    /// Initializes all internal state information and data in this instance of System.Text.EncoderReplacementFallbackBuffer.
+    /// </summary>
+    [SecuritySafeCritical]
+    public override void Reset()
+    {
+        if (Buffer.Any())
+        {
+            Buffer = new ArraySegment<char>(Buffer.Array, 0, Buffer.Array.Length);
         }
     }
 }
